+++ x-pack/solutions/security/plugins/security_solution/public/entity_analytics/components/privileged_user_monitoring/components/privileged_access_detection/pad_chart/index.tsx
@@ -6,11 +6,14 @@
  */
 import React from 'react';
 import { EuiFlexGroup } from '@elastic/eui';
+import { useGlobalTime } from '../../../../../../common/containers/use_global_time';
 import { PrivilegedAccessDetectionSeverityFilter } from './pad_chart_severity_filter';
 import { usePrivilegedAccessDetectionAnomaliesQuery } from './hooks/pad_query_hooks';
 import { useAnomalyBands } from './pad_anomaly_bands';
 import { UserNameList } from './pad_user_name_list';
 import { PrivilegedAccessDetectionHeatmap } from './pad_heatmap';
+import { useQueryInspector } from '../../../../../../common/components/page/manage_query';
+import { PRIVILEGED_ACCESS_DETECTIONS_QUERY_ID } from '..';
 
 export interface PrivilegedAccessDetectionChartProps {
   jobIds: string[];
@@ -21,12 +24,25 @@ export const PrivilegedAccessDetectionChart: React.FC<PrivilegedAccessDetectionC
   jobIds,
   spaceId,
 }) => {
+  const { deleteQuery, setQuery } = useGlobalTime();
+
   const { bands, toggleHiddenBand } = useAnomalyBands();
 
-  const { data, isLoading, isError } = usePrivilegedAccessDetectionAnomaliesQuery({
-    jobIds,
-    anomalyBands: bands,
-    spaceId,
+  const { data, isLoading, isError, inspect, refetch } = usePrivilegedAccessDetectionAnomaliesQuery(
+    {
+      jobIds,
+      anomalyBands: bands,
+      spaceId,
+    }
+  );
+
+  useQueryInspector({
+    deleteQuery,
+    inspect,
+    refetch,
+    setQuery,
+    queryId: PRIVILEGED_ACCESS_DETECTIONS_QUERY_ID,
+    loading: isLoading,
   });
 
   return (
