+++ x-pack/solutions/security/plugins/security_solution/public/entity_analytics/components/privileged_user_monitoring/components/privileged_users_table/index.tsx
@@ -4,7 +4,7 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-import React, { useMemo, useState } from 'react';
+import React, { useState } from 'react';
 import {
   useEuiTheme,
   EuiPanel,
@@ -19,26 +19,18 @@ import {
   EuiCallOut,
 } from '@elastic/eui';
 import { css } from '@emotion/react';
-import { take } from 'lodash/fp';
 import { FormattedMessage } from '@kbn/i18n-react';
 import { useExpandableFlyoutApi } from '@kbn/expandable-flyout';
-import { useQuery } from '@tanstack/react-query';
-import { getESQLResults } from '@kbn/esql-utils';
 import { i18n } from '@kbn/i18n';
+import { InspectButtonContainer } from '../../../../../common/components/inspect';
+import { useGlobalTime } from '../../../../../common/containers/use_global_time';
+import { useQueryInspector } from '../../../../../common/components/page/manage_query';
 import { useQueryToggle } from '../../../../../common/containers/query_toggle';
-import { useRiskScore } from '../../../../api/hooks/use_risk_score';
-import type { TableItemType } from './types';
-import { getPrivilegedUsersQuery } from './esql_source_query';
 import { UserPanelKey } from '../../../../../flyout/entity_details/shared/constants';
-import type { EntityRiskScore } from '../../../../../../common/search_strategy';
-import { buildEntityNameFilter, EntityType } from '../../../../../../common/search_strategy';
+
 import { buildPrivilegedUsersTableColumns } from './columns';
-import { esqlResponseToRecords } from '../../../../../common/utils/esql';
-import { useKibana } from '../../../../../common/lib/kibana';
-import { useGlobalFilterQuery } from '../../../../../common/hooks/use_global_filter_query';
 import { HeaderSection } from '../../../../../common/components/header_section';
-import { useAssetCriticalityFetchList } from '../../../asset_criticality/use_asset_criticality';
-import type { CriticalityLevelWithUnassigned } from '../../../../../../common/entity_analytics/asset_criticality/types';
+import { usePrivilegedUsersTableData } from './hooks';
 
 export const DEFAULT_PAGE_SIZE = 10;
 
@@ -68,230 +60,118 @@ const useOpenUserFlyout = () => {
   };
 };
 
-interface RiskScoresByUserName {
-  [key: string]: EntityRiskScore<EntityType.user>;
-}
-
-interface AssetCriticalityByUserName {
-  [key: string]: CriticalityLevelWithUnassigned;
-}
-
 export const PrivilegedUsersTable: React.FC<{ spaceId: string }> = ({ spaceId }) => {
+  const { setQuery, deleteQuery } = useGlobalTime();
   const { toggleStatus, setToggleStatus } = useQueryToggle(PRIVILEGED_USERS_TABLE_QUERY_ID);
 
   const { euiTheme } = useEuiTheme();
-
-  const { data } = useKibana().services;
 
   const openUserFlyout = useOpenUserFlyout();
 
   const columns = buildPrivilegedUsersTableColumns(openUserFlyout, euiTheme);
   const [currentPage, setCurrentPage] = useState<number>(1);
+  const { visibleRecords, isLoading, hasError, refetch, inspect, hasNextPage } =
+    usePrivilegedUsersTableData(spaceId, currentPage, toggleStatus);
 
-  const privilegedUsersTableQuery = getPrivilegedUsersQuery(
-    spaceId,
-    currentPage * DEFAULT_PAGE_SIZE + 1 // we add 1 so that we know if there are more results to show
-  );
-
-  const { filterQuery: filterQueryWithoutTimerange } = useGlobalFilterQuery();
-
-  const {
-    data: result,
-    isLoading: loadingPrivilegedUsers,
-    isError: privilegedUsersError,
-  } = useQuery({
-    queryKey: ['privileged-users-table', privilegedUsersTableQuery, filterQueryWithoutTimerange],
-    enabled: toggleStatus,
-    queryFn: async () => {
-      return getESQLResults({
-        esqlQuery: privilegedUsersTableQuery,
-        search: data.search.search,
-        filter: filterQueryWithoutTimerange,
-      });
-    },
+  useQueryInspector({
+    deleteQuery,
+    inspect,
+    refetch,
+    setQuery,
+    queryId: PRIVILEGED_USERS_TABLE_QUERY_ID,
+    loading: isLoading,
   });
 
-  const records = useMemo(() => esqlResponseToRecords<TableItemType>(result?.response), [result]);
-
-  const nameFilterQuery = useMemo(() => {
-    const userNames = records.map((user) => user['user.name']);
-    return buildEntityNameFilter(EntityType.user, userNames);
-  }, [records]);
-
-  const {
-    data: riskScoreData,
-    error: riskScoreError,
-    loading: loadingRiskScore,
-    hasEngineBeenInstalled: hasRiskEngineBeenInstalled,
-  } = useRiskScore<EntityType.user>({
-    riskEntity: EntityType.user,
-    filterQuery: nameFilterQuery,
-    onlyLatest: true,
-    pagination: {
-      cursorStart: 0,
-      querySize: records.length,
-    },
-    skip: nameFilterQuery === undefined || records.length === 0,
-  });
-
-  const riskScores = riskScoreData && riskScoreData.length > 0 ? riskScoreData : [];
-
-  const riskScoreByUserName: RiskScoresByUserName = Object.fromEntries(
-    riskScores.map((riskScore) => [riskScore.user.name, riskScore])
-  );
-  const {
-    data: assetCriticalityData,
-    isError: assetCriticalityError,
-    isLoading: loadingAssetCriticality,
-  } = useAssetCriticalityFetchList({
-    idField: 'user.name',
-    idValues: records.map((user) => user['user.name']),
-    skip: !toggleStatus || records.length === 0,
-  });
-  const assetCriticalityRecords =
-    assetCriticalityData && assetCriticalityData.records.length > 0
-      ? assetCriticalityData.records
-      : [];
-
-  const assetCriticalityByUserName: AssetCriticalityByUserName = Object.fromEntries(
-    assetCriticalityRecords.map((assetCriticalityRecord) => [
-      assetCriticalityRecord.id_value,
-      assetCriticalityRecord.criticality_level,
-    ])
-  );
-
-  const enrichedRecords: TableItemType[] = useMemo(
-    () =>
-      records.map((record, index) => {
-        let enrichedFields = {};
-
-        const riskScore: EntityRiskScore<EntityType.user> | undefined =
-          riskScoreByUserName[record['user.name']];
-        if (riskScore) {
-          enrichedFields = {
-            ...enrichedFields,
-            risk_score: riskScore.user.risk.calculated_score_norm,
-            risk_level: riskScore.user.risk.calculated_level,
-          };
-        }
-
-        const assetCriticality: CriticalityLevelWithUnassigned | undefined =
-          assetCriticalityByUserName[record['user.name']];
-
-        if (assetCriticality) {
-          enrichedFields = {
-            ...enrichedFields,
-            criticality_level: assetCriticality,
-          };
-        }
-
-        return {
-          ...record,
-          ...enrichedFields,
-        };
-      }),
-    [records, riskScoreByUserName, assetCriticalityByUserName]
-  );
-
-  const isLoading =
-    loadingPrivilegedUsers || (records.length > 0 && (loadingRiskScore || loadingAssetCriticality));
-  const visibleRecords = take(currentPage * DEFAULT_PAGE_SIZE, enrichedRecords);
   return (
-    <EuiPanel hasBorder hasShadow={false} data-test-subj="privileged-users-table-panel">
-      <HeaderSection
-        toggleStatus={toggleStatus}
-        toggleQuery={setToggleStatus}
-        id={PRIVILEGED_USERS_TABLE_QUERY_ID}
-        showInspectButton={false}
-        title={TITLE}
-        titleSize="s"
-        outerDirection="column"
-        hideSubtitle
-      />
-      {(privilegedUsersError ||
-        (hasRiskEngineBeenInstalled && riskScoreError) ||
-        assetCriticalityError) && (
-        <EuiCallOut
-          title={i18n.translate(
-            'xpack.securitySolution.entityAnalytics.privilegedUserMonitoring.privilegedUsersTable.error',
-            {
-              defaultMessage:
-                'There was an error retrieving privileged users. Results may be incomplete.',
-            }
-          )}
-          color="danger"
-          iconType="error"
+    <InspectButtonContainer>
+      <EuiPanel hasBorder hasShadow={false} data-test-subj="privileged-users-table-panel">
+        <HeaderSection
+          toggleStatus={toggleStatus}
+          toggleQuery={setToggleStatus}
+          id={PRIVILEGED_USERS_TABLE_QUERY_ID}
+          title={TITLE}
+          titleSize="s"
+          outerDirection="column"
+          hideSubtitle
         />
-      )}
-      {toggleStatus && (
-        <EuiFlexGroup direction="column" gutterSize="s">
-          <EuiFlexItem>
-            {isLoading &&
-              (loadingPrivilegedUsers || loadingRiskScore || loadingAssetCriticality) && (
-                <EuiProgress size="xs" color="accent" />
-              )}
-          </EuiFlexItem>
-          {records.length > 0 ? (
-            <>
-              <EuiText size={'s'}>
-                <FormattedMessage
-                  id="xpack.securitySolution.entityAnalytics.privilegedUserMonitoring.privilegedUsersTable.showing"
-                  defaultMessage="Showing "
+        {hasError && (
+          <EuiCallOut
+            title={i18n.translate(
+              'xpack.securitySolution.entityAnalytics.privilegedUserMonitoring.privilegedUsersTable.error',
+              {
+                defaultMessage:
+                  'There was an error retrieving privileged users. Results may be incomplete.',
+              }
+            )}
+            color="danger"
+            iconType="error"
+          />
+        )}
+        {toggleStatus && (
+          <EuiFlexGroup direction="column" gutterSize="s">
+            <EuiFlexItem>{isLoading && <EuiProgress size="xs" color="accent" />}</EuiFlexItem>
+            {visibleRecords.length > 0 ? (
+              <>
+                <EuiText size={'s'}>
+                  <FormattedMessage
+                    id="xpack.securitySolution.entityAnalytics.privilegedUserMonitoring.privilegedUsersTable.showing"
+                    defaultMessage="Showing "
+                  />
+                  <span
+                    css={css`
+                      font-weight: ${euiTheme.font.weight.bold};
+                    `}
+                  >
+                    <FormattedMessage
+                      id="xpack.securitySolution.entityAnalytics.privilegedUserMonitoring.privilegedUsersTable.countOfUsers"
+                      defaultMessage="{count} privileged {count, plural, one {user} other {users}}"
+                      values={{ count: visibleRecords.length }}
+                    />
+                  </span>
+                </EuiText>
+                <EuiSpacer size="s" />
+                <EuiHorizontalRule margin="none" css={{ height: 2 }} />
+                <EuiBasicTable
+                  id={PRIVILEGED_USERS_TABLE_QUERY_ID}
+                  loading={isLoading}
+                  items={visibleRecords || []}
+                  columns={columns}
                 />
-                <span
-                  css={css`
-                    font-weight: ${euiTheme.font.weight.bold};
-                  `}
+              </>
+            ) : (
+              !isLoading && (
+                <EuiText size="s" color="subdued" textAlign="center">
+                  <FormattedMessage
+                    id="xpack.securitySolution.entityAnalytics.privilegedUserMonitoring.privilegedUsersTable.noData"
+                    defaultMessage="No privileged users found"
+                  />
+                </EuiText>
+              )
+            )}
+            {hasNextPage && (
+              <EuiFlexItem grow={false}>
+                <EuiButtonEmpty
+                  isLoading={isLoading}
+                  onClick={() => {
+                    setCurrentPage((page) => page + 1);
+                  }}
+                  flush="right"
+                  color="primary"
+                  size="s"
+                  iconType="sortDown"
+                  iconSide="right"
+                  iconSize="s"
                 >
                   <FormattedMessage
-                    id="xpack.securitySolution.entityAnalytics.privilegedUserMonitoring.privilegedUsersTable.countOfUsers"
-                    defaultMessage="{count} privileged {count, plural, one {user} other {users}}"
-                    values={{ count: visibleRecords.length }}
+                    id="xpack.securitySolution.privilegedUserMonitoring.showMore"
+                    defaultMessage="Show more"
                   />
-                </span>
-              </EuiText>
-              <EuiSpacer size="s" />
-              <EuiHorizontalRule margin="none" style={{ height: 2 }} />
-              <EuiBasicTable
-                id={PRIVILEGED_USERS_TABLE_QUERY_ID}
-                loading={isLoading}
-                items={visibleRecords || []}
-                columns={columns}
-              />
-            </>
-          ) : (
-            !isLoading && (
-              <EuiText size="s" color="subdued" textAlign="center">
-                <FormattedMessage
-                  id="xpack.securitySolution.entityAnalytics.privilegedUserMonitoring.privilegedUsersTable.noData"
-                  defaultMessage="No privileged users found"
-                />
-              </EuiText>
-            )
-          )}
-          {records.length > currentPage * DEFAULT_PAGE_SIZE && (
-            <EuiFlexItem grow={false}>
-              <EuiButtonEmpty
-                isLoading={isLoading}
-                onClick={() => {
-                  setCurrentPage((page) => page + 1);
-                }}
-                flush="right"
-                color="primary"
-                size="s"
-                iconType="sortDown"
-                iconSide="right"
-                iconSize="s"
-              >
-                <FormattedMessage
-                  id="xpack.securitySolution.privilegedUserMonitoring.showMore"
-                  defaultMessage="Show more"
-                />
-              </EuiButtonEmpty>
-            </EuiFlexItem>
-          )}
-        </EuiFlexGroup>
-      )}
-    </EuiPanel>
+                </EuiButtonEmpty>
+              </EuiFlexItem>
+            )}
+          </EuiFlexGroup>
+        )}
+      </EuiPanel>
+    </InspectButtonContainer>
   );
 };
